+++ hs/src/HelVM/HelMA/Automata/ETA/Evaluator.hs
@@ -9,6 +9,7 @@ import           HelVM.HelMA.Automata.ETA.Optimizer
 import           HelVM.HelMA.Automata.ETA.Symbol
 import           HelVM.HelMA.Automata.ETA.Token
 
+import           HelVM.HelMA.Automaton.API.AutoParams
 import           HelVM.HelMA.Automaton.API.EvalParams
 import           HelVM.HelMA.Automaton.API.IOTypes
 
@@ -17,12 +18,11 @@ import qualified HelVM.HelMA.Automaton.Automaton          as Automaton
 import           HelVM.HelMA.Automaton.IO.AutomatonIO
 import           HelVM.HelMA.Automaton.IO.BusinessIO
 
-import           HelVM.HelMA.Automaton.Loop
-
 import           HelVM.HelMA.Automaton.Types.DumpType
 import           HelVM.HelMA.Automaton.Types.StackType
 
 import           HelVM.HelIO.Collections.SList            as SList
+
 
 import           Prelude                                  hiding (divMod)
 
@@ -31,21 +31,21 @@ import qualified Data.Sequence                            as Seq
 import           HelVM.HelMA.Automata.ETA.API.ETAImplType
 
 simpleEval :: BIO m => (ETAImplType , Source , StackType) -> m ()
-simpleEval (c , s , t) = evalSource c s t testMaybeLimit Pretty
+simpleEval (c , s , t) = evalSource c s t verySimpleAutoParams
 
 ----
 
 evalParams :: BIO m => ETAImplType -> EvalParams -> m ()
-evalParams e p = evalSource e (source p) (stackAutoOptions p) Nothing (dumpAutoOptions p)
+evalParams e p = evalSource e (source p) (stackAutoOptions p) (autoParams p)
 
-evalSource :: (AutomatonIO Symbol m) => ETAImplType -> Source -> StackType -> LimitMaybe -> DumpType -> m ()
+evalSource :: (AutomatonIO Symbol m) => ETAImplType -> Source -> StackType -> AutoParams -> m ()
 evalSource etaImplType source = evalTL etaImplType (tokenize source)
 
-evalTL :: (AutomatonIO Symbol m) => ETAImplType -> TokenList -> StackType -> LimitMaybe -> DumpType -> m ()
+evalTL :: (AutomatonIO Symbol m) => ETAImplType -> TokenList -> StackType -> AutoParams -> m ()
 evalTL c tl ListStackType  = eval c tl []
 evalTL c tl SeqStackType   = eval c tl Seq.empty
 evalTL c tl SListStackType = eval c tl SList.sListEmpty
 
-eval :: (SAutomatonIO Symbol s m) => ETAImplType -> TokenList -> s -> LimitMaybe -> DumpType -> m ()
-eval Original tl s limit dt = logDump dt =<< run limit (newAutomaton tl s)
-eval Fast     tl s limit dt = logDump dt =<< (Automaton.run limit . Automaton.flippedNewAutomaton (s , [])) =<< optimize tl
+eval :: (SAutomatonIO Symbol s m) => ETAImplType -> TokenList -> s -> AutoParams -> m ()
+eval Fast     tl s (AutoParams c limit dt) = logDump dt =<< (Automaton.run c limit . Automaton.flippedNewAutomaton (s , [])) =<< optimize tl
+eval Original tl s (AutoParams _ limit dt) = logDump dt =<< run limit (newAutomaton tl s)