+++ hs/test/HelVM/HelMA/Automaton/API/AutoOptionsExtra.hs
@@ -3,7 +3,6 @@ module HelVM.HelMA.Automaton.API.AutoOptionsExtra where
 import           HelVM.HelMA.Automaton.API.AutoOptions
 
 import           HelVM.HelMA.Automaton.Types.CellType
-import           HelVM.HelMA.Automaton.Types.DumpType
 import           HelVM.HelMA.Automaton.Types.IntCellType
 import           HelVM.HelMA.Automaton.Types.RAMType
 import           HelVM.HelMA.Automaton.Types.StackType
@@ -14,5 +13,4 @@ defaultAutoOptions = AutoOptions
   , stack    = defaultStackType
   , cell     = defaultCellType
   , intCell  = defaultIntCellType
-  , dumpType = defaultDumpType
   }